Refactor: remove settings.py
- weekly build time slot selection moved to
  classmethod in Distro()

- ABCFG_DICT moved to abconfig.py
